requirements.txt
bs4
chromadb
gpt4all
httptools
ipykernel
ipywidgets
langchain
langchainhub
llama-cpp-python
llama-index
matplotlib
pypdf
python-dotenv
rank-bm25
sentence-transformers
textract
watchfiles
websockets